+++ bootstrap-frontend-play-30/src/main/scala/uk/gov/hmrc/play/bootstrap/frontend/filters/crypto/ApplicationCryptoProvider.scala
@@ -16,16 +16,65 @@
 
 package uk.gov.hmrc.play.bootstrap.frontend.filters.crypto
 
-import javax.inject.{Inject, Provider}
 import play.api.Configuration
-import uk.gov.hmrc.crypto.ApplicationCrypto
+import uk.gov.hmrc.crypto.SymmetricCryptoFactory
+
+import javax.inject.{Inject, Provider}
 
 class ApplicationCryptoProvider @Inject()(
   configuration: Configuration
 ) extends Provider[ApplicationCrypto] {
 
-  private val crypto = new ApplicationCrypto(configuration.underlying)
+  private val crypto =
+    new ApplicationCrypto(configuration)
   crypto.verifyConfiguration()
 
-  def get(): ApplicationCrypto = crypto
+  def get(): ApplicationCrypto =
+    crypto
+}
+
+class ApplicationCrypto @Inject()(configuration: Configuration) {
+
+  /** Used to encrypt the cookie.
+    * It is shared by all services.
+    * This is a platform key, and should not be used for any other use-case since it may be rotated at any time.
+    */
+  lazy val SessionCookieCrypto =
+    SymmetricCryptoFactory.aesGcmCryptoFromConfig(baseConfigKey = "cookie.encryption", configuration.underlying)
+
+  /** Used for SSO with with the Portal.
+    * It is shared by all services.
+    * This is a platform key, and should not be used for any other use-case since it may be rotated at any time.
+    */
+  lazy val SsoPayloadCrypto =
+    SymmetricCryptoFactory.aesCryptoFromConfig(baseConfigKey = "sso.encryption", configuration.underlying)
+
+  /** Used to encrypt query parameters. It can be used for path parameters and json payloads too.
+    * It is shared by all services.
+    * This is a platform key, and should not be used for any other use-case since it may be rotated at any time.
+    */
+  lazy val QueryParameterCrypto =
+    SymmetricCryptoFactory.aesCryptoFromConfig(baseConfigKey = "queryParameter.encryption", configuration.underlying)
+
+  def verifyConfiguration(): Unit = {
+    SessionCookieCrypto
+    QueryParameterCrypto
+    SsoPayloadCrypto
+  }
+}
+
+class DeprecatedApplicationCryptoProvider @Inject()(
+  configuration: Configuration
+) extends Provider[uk.gov.hmrc.crypto.ApplicationCrypto] {
+
+  private val logger = play.api.Logger(getClass)
+
+  private val crypto =
+    new uk.gov.hmrc.crypto.ApplicationCrypto(configuration.underlying)
+  crypto.verifyConfiguration()
+
+  def get(): uk.gov.hmrc.crypto.ApplicationCrypto = {
+    logger.warn("uk.gov.hmrc.crypto.ApplicationCrypto is deprecated. Inject uk.gov.hmrc.play.bootstrap.frontend.filters.crypto.AppliationCrypto instead.")
+    crypto
+  }
 }